Create Research Objectives
This research seeks to explore where law and technology meet. The first step is to identify the global fortune hundred (100) companies for the year ending 2022 and compile their geographical dispersion. The focus is to establish processing concerns in the governance policies and then use the appropriate programing language to purge or develop rules, triggers and algorithms to solve such issues. This is to help technology experts and non-technology experts such as legal practitioners, management etc. to formulate and implement governance policies with ease and clarity. This will also eliminate all ambiguities in the governance policies for corporate bodies.